- Moved Bank Account Totals Component to shared components.
- Added GET API call and used in ListOfAccounts component in order to display bank accounts.
