+++ personalfinanceapp/src/pages/SharedComponents/BankAccountTotals.jsx
@@ -11,6 +11,18 @@ const BankAccountTotals = ({ accounts, pending, label }) => {
     return <div>Loading Bank Account Data ...</div>
   }
 
+  if (label === "Bank Accounts"){
+    return(
+     <div className="grid md:grid-cols-2 lg:grid-cols-3 gap-4">
+        {accounts.map((account) => (
+         <div key={account.id} className="eachBankAccount">
+            <h3 className="text-lg font-medium smallerMargin">{account.name}</h3>
+            <p className="text-xl font-bold smallerMargin">{currencyFormat.format(account.balance)}</p>
+         </div>
+        ))}
+      </div>
+    );
+  }
   if(label != "Total After Pending")
   { 
    return (
